atualizações, responseentity, exceptions e padrão DTO

+++ src/main/java/com/example/usuario/api/service/UsuarioService.java
@@ -2,6 +2,7 @@ package com.example.usuario.api.service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -11,6 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 
 import com.example.usuario.api.usuarioDTO;
+import com.example.usuario.api.exceptionahandler.DataIntegratedViolationException;
 import com.example.usuario.api.model.Usuario;
 import com.example.usuario.api.repository.UsuarioRepository;
 
@@ -21,16 +23,29 @@ public class UsuarioService {
 	@Autowired
 	private UsuarioRepository usuarioRepository;
 	
-	public List<Usuario> listarTodos() {
-		return usuarioRepository.findAll();
+	public List <usuarioDTO> listarTodos() {
+		List<Usuario> lista = usuarioRepository.findAll();
+		List<usuarioDTO> listaDTO = lista.stream().map(obj -> new usuarioDTO(obj)).collect(Collectors.toList());
+		return listaDTO;
 	}
 	
 	public usuarioDTO listarPorID(Long id) {
-		Usuario entity = usuarioRepository.findById(id).get();
-		usuarioDTO objDTO = new usuarioDTO(entity);
-		return objDTO;	
+		try {
+			Optional<Usuario> resposta = usuarioRepository.findById(id);
+			if (resposta.isPresent()) {
+				Usuario entity = usuarioRepository.findById(id).get();
+				usuarioDTO objDTO = new usuarioDTO(entity);
+				return objDTO;
+			} else {
+				throw new DataIntegratedViolationException("Usuário não encontrado");
+			}
+			
+		} catch (Exception e) {
+			throw new DataIntegratedViolationException("Falha ao deletar o usuário", e);
+		}
 		
 	}
+
 	
 	public usuarioDTO cadastrarUsuario(@RequestBody Usuario usuario) {
 		Usuario entity = usuarioRepository.save(usuario);
@@ -39,20 +54,33 @@ public class UsuarioService {
 	}
 	
 	public void deletaUsuario(@PathVariable Long id) {
-	usuarioRepository.deleteById(id);
+		try {	
+			Optional<Usuario> resposta = usuarioRepository.findById(id);
+			if (resposta.isPresent()) {
+				usuarioRepository.deleteById(id);
+			} else {
+				throw new DataIntegratedViolationException("Usuário não encontrado");
+			}
+		} catch (Exception e) {
+			throw new DataIntegratedViolationException("Falha ao deletar o usuário", e);		
+		}
 	}
 	
-	public ResponseEntity<?> atualizarUsuario(@RequestBody Usuario usuario) {
-		
-		if (usuario.getId() == null) {
-			return new ResponseEntity<String>("Id não foi informado para atualização", HttpStatus.NOT_FOUND);
-			
+	public usuarioDTO atualizarUsuario(@RequestBody Usuario usuario) {
+		try {
+			Optional<Usuario> resposta = usuarioRepository.findById(usuario.getId());
+			if (resposta.isPresent()) {
+				Usuario entity = usuarioRepository.save(usuario);
+				usuarioDTO objDTO = new usuarioDTO(entity);
+				return objDTO;
+			} else {
+				throw new DataIntegratedViolationException("Usuário não encontrado");
+			}
+		} catch (Exception e) {
+			throw new DataIntegratedViolationException("Falha ao deletar o usuário", e);
 		}
-		
-		Usuario user = usuarioRepository.saveAndFlush(usuario);
-		
-		return new ResponseEntity<Usuario>(usuario, HttpStatus.OK);
-		
+			
 	}
-
+		
 }
+
